+++ ivy/functional/frontends/numpy/mathematical_functions/__init__.py
@@ -23,5 +23,3 @@ from . import extrema_finding
 from .extrema_finding import *
 from . import miscellaneous
 from .miscellaneous import *
-from . import lcm_and_gcd
-from .lcm_and_gcd import *
